backend/requirements.txt
django>=5.2,<5.3
djangorestframework>=3.14
django-cors-headers>=4.3
mssql-django>=1.4
Pillow>=10.0
gunicorn>=21.2
